Initial commit
This adds the basic structure that I wanted to see and ease the
setup of new computers/users in future.

The plan is to switch users every month and clean up the previous ones
so that this repo does not stagnate and keeps getting new configs and
features.

Next up:
- Ratpoison
- Tsundere